feat: risk level in release check
