+++ proto/gen/grpc-doc/v1/index.html
@@ -26061,6 +26061,13 @@ Should be a database. Format: instances/{instance}/databases/{database} </p></td
                   <td><p>The count of affected rows of the statement on the target. </p></td>
                 </tr>
               
+                <tr>
+                  <td>risk_level</td>
+                  <td><a href="#bytebase.v1.CheckReleaseResponse.RiskLevel">CheckReleaseResponse.RiskLevel</a></td>
+                  <td></td>
+                  <td><p>The risk level of the statement on the target. </p></td>
+                </tr>
+              
             </tbody>
           </table>
